:sparkles: Apollo Client added to frontend and products base data querys done
In order to easily access to GraphQL sentences, was added to frontend as a dev dependency graphql-codegen, this allows us to simplify query process. Also was added Apollo context, so, custom hooks to query data are running. In backend all current work to frontend is done. TODO: Add shopping list and user authentication support to backend.